refactor: introduce packet building utilities
