--- src/handlers/handlerf.rs
@@ -1,175 +1,0 @@
-use bytes::{Buf, BufMut, BytesMut};
-use std::error::Error;
-use std::sync::Arc;
-use std::time::Instant;
-use tracing::{debug, info, warn};
-
-use crate::kaillera::message_types as msg;
-use crate::*;
-
-pub async fn handle_message(
-    message: kaillera::protocol::ParsedMessage,
-    src: &std::net::SocketAddr,
-    state: Arc<AppState>,
-) -> Result<(), Box<dyn Error>> {
-    match message.message_type {
-        msg::USER_QUIT => user_quit::handle_user_quit(message, src, state).await?,
-        msg::USER_LOGIN => user_login::handle_user_login(message, src, state).await?,
-        // msg::SERVER_STATUS => handle_server_status(src, state).await?,
-        // msg::SERVER_TO_CLIENT_ACK => handle_server_to_client_ack(message, src, state).await?,
-        msg::CLIENT_TO_SERVER_ACK => handle_client_to_server_ack(src, state).await?,
-        msg::GLOBAL_CHAT => global_chat::handle_global_chat(message, src, state).await?,
-        msg::GAME_CHAT => game_chat::handle_game_chat(message, src, state).await?,
-        msg::CLIENT_KEEP_ALIVE => handle_client_keep_alive(message, src).await?,
-        msg::CREATE_GAME => create_game::handle_create_game(message, src, state).await?,
-        msg::QUIT_GAME => handlers::quit_game::handle_quit_game(message.data, src, state).await?,
-        msg::JOIN_GAME => join_game::handle_join_game(message, src, state).await?,
-        msg::KICK_USER => {
-            kick_user::handle_kick_user(message, src, state).await?;
-        }
-        msg::START_GAME => start_game::handle_start_game(message, src, state).await?,
-        msg::GAME_DATA => {
-            debug!(
-                message_type = msg::message_type_name(message.message_type),
-                "Game sync request received"
-            );
-            game_data::handle_game_data(message, src, state).await?;
-        }
-        msg::GAME_CACHE => game_cache::handle_game_cache(message, src, state).await?,
-        msg::DROP_GAME => drop_game::handle_drop_game(message, src, state).await?,
-        msg::READY_TO_PLAY => handle_ready_to_play_signal(message, src, state).await?,
-
-        _ => {
-            warn!(
-                message_type = msg::message_type_name(message.message_type),
-                "Unknown message type received"
-            );
-        }
-    }
-    Ok(())
-}
-
-pub async fn handle_client_to_server_ack(
-    src: &std::net::SocketAddr,
-    state: Arc<AppState>,
-) -> Result<(), Box<dyn Error>> {
-    // Client to Server ACK [0x06]
-    // Calculate ping and update ack count
-    let ack_count = state
-        .update_client::<_, u16, Box<dyn Error>>(src, |client_info| {
-            if let Some(last_ping_time) = client_info.last_ping_time {
-                let ping = last_ping_time.elapsed().as_millis() as u32;
-                client_info.ping = ping;
-                client_info.last_ping_time = Some(Instant::now());
-                client_info.ack_count += 1;
-            }
-            Ok(client_info.ack_count)
-        })
-        .await?;
-
-    if ack_count >= 3 {
-        let data = util::make_server_status(src, &state).await?;
-        util::send_packet(&state, src, msg::SERVER_STATUS, data).await?;
-
-        let data = util::make_user_joined(src, &state).await?;
-        util::broadcast_packet(&state, msg::USER_JOINED, data).await?;
-
-        let data = util::make_server_information()?;
-        util::send_packet(&state, src, msg::SERVER_INFORMATION, data).await?;
-    } else {
-        // Server notification creation
-        let mut data = BytesMut::new();
-        data.put_u8(0);
-        data.put_u32_le(0);
-        data.put_u32_le(1);
-        data.put_u32_le(2);
-        data.put_u32_le(3);
-        util::send_packet(&state, src, msg::SERVER_TO_CLIENT_ACK, data.to_vec()).await?;
-    }
-
-    Ok(())
-}
-
-pub async fn handle_client_keep_alive(
-    _message: kaillera::protocol::ParsedMessage,
-    _src: &std::net::SocketAddr,
-) -> Result<(), Box<dyn Error>> {
-    // No additional handling needed
-    Ok(())
-}
-
-pub async fn handle_ready_to_play_signal(
-    message: kaillera::protocol::ParsedMessage,
-    src: &std::net::SocketAddr,
-    state: Arc<AppState>,
-) -> Result<(), Box<dyn Error>> {
-    debug!("Ready to play signal received");
-    let mut buf = BytesMut::from(&message.data[..]);
-    let _ = buf.get_u8(); // Empty String
-
-    state
-        .update_client::<_, (), Box<dyn Error>>(src, |client_info| {
-            client_info.player_status = PLAYER_STATUS_NET_SYNC; // Ready to play
-            Ok(())
-        })
-        .await?;
-
-    let game_info_clone = util::fetch_game_info(src, &state).await?;
-
-    // Update game status
-    {
-        let status_data = util::make_update_game_status(&game_info_clone)?;
-        util::broadcast_packet(&state, msg::UPDATE_GAME_STATUS, status_data).await?;
-    }
-
-    // Check if all users are ready
-    let all_user_ready_to_signal = {
-        let addr_map = state.clients_by_addr.read().await;
-        let id_map = state.clients_by_id.read().await;
-
-        let all_ready = game_info_clone.players.iter().all(|player| {
-            if let Some(session_id) = addr_map.get(&player.addr) {
-                if let Some(client_info) = id_map.get(session_id) {
-                    debug!(
-                        { fields::ADDR } = %player.addr,
-                        player_status = client_info.player_status,
-                        "Checking player status"
-                    );
-                    return client_info.player_status == PLAYER_STATUS_NET_SYNC;
-                }
-            }
-            debug!(
-                { fields::ADDR } = %player.addr,
-                "Client info not found"
-            );
-            false
-        });
-        all_ready
-    };
-
-    // If all ready, update all players' status
-    if all_user_ready_to_signal {
-        for player in &game_info_clone.players {
-            let _ = state
-                .update_client::<_, (), Box<dyn Error>>(&player.addr, |client_info| {
-                    client_info.player_status = PLAYER_STATUS_PLAYING;
-                    Ok(())
-                })
-                .await;
-        }
-    }
-
-    // Send ready to play signal notification
-    if all_user_ready_to_signal {
-        info!(
-            { fields::PLAYER_COUNT } = game_info_clone.players.len(),
-            "All users ready to signal - starting game"
-        );
-        for player in &game_info_clone.players {
-            let mut data = BytesMut::new();
-            data.put_u8(0);
-            util::send_packet(&state, &player.addr, msg::READY_TO_PLAY, data.to_vec()).await?;
-        }
-    }
-    Ok(())
-}
