+++ src/packet_util.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::*;
-use bytes::{Buf, BytesMut};
+use bytes::{Buf, BufMut, BytesMut};
 use tokio::io;
 use tokio::net::UdpSocket;
 use tracing::{debug, error};
@@ -84,4 +84,128 @@ pub fn parse_packet(data: &[u8]) -> Result<Vec<kaillera::protocol::ParsedMessage
     }
 
     Ok(messages)
+}
+
+// Packet building utilities
+
+/// Appends a string with null terminator to the buffer
+pub fn put_string_with_null(buf: &mut BytesMut, s: &str) {
+    buf.put(s.as_bytes());
+    buf.put_u8(0);
+}
+
+/// Appends an empty string (just null terminator) to the buffer
+pub fn put_empty_string(buf: &mut BytesMut) {
+    buf.put_u8(0);
+}
+
+/// Appends multiple strings with null terminators to the buffer
+pub fn put_strings_with_null(buf: &mut BytesMut, strings: &[&str]) {
+    for s in strings {
+        put_string_with_null(buf, s);
+    }
+}
+
+// Packet builder functions - return Vec<u8> for packet data
+
+/// Builds a START_GAME packet
+/// Format: Empty String [00], Frame Delay (2B), Player Number (1B), Total Players (1B)
+pub fn build_start_game_packet(frame_delay: u16, player_number: u8, total_players: u8) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_empty_string(&mut data);
+    data.put_u16_le(frame_delay);
+    data.put_u8(player_number);
+    data.put_u8(total_players);
+    data.to_vec()
+}
+
+/// Builds a GAME_CHAT packet
+/// Format: Username (NB), Message (NB)
+pub fn build_game_chat_packet(username: &str, message: &str) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_string_with_null(&mut data, username);
+    put_string_with_null(&mut data, message);
+    data.to_vec()
+}
+
+/// Builds a GLOBAL_CHAT packet
+/// Format: Username (NB), Message (NB)
+pub fn build_global_chat_packet(username: &str, message: &str) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_string_with_null(&mut data, username);
+    put_string_with_null(&mut data, message);
+    data.to_vec()
+}
+
+/// Builds a USER_QUIT packet
+/// Format: Username (NB), UserID (2B), Message (NB)
+pub fn build_user_quit_packet(username: &str, user_id: u16, message: &str) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_string_with_null(&mut data, username);
+    data.put_u16_le(user_id);
+    put_string_with_null(&mut data, message);
+    data.to_vec()
+}
+
+/// Builds a QUIT_GAME packet
+/// Format: Username (NB), UserID (2B)
+pub fn build_quit_game_packet(username: &str, user_id: u16) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_string_with_null(&mut data, username);
+    data.put_u16_le(user_id);
+    data.to_vec()
+}
+
+/// Builds a CLOSE_GAME packet
+/// Format: Empty String [00], GameID (4B)
+pub fn build_close_game_packet(game_id: u32) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_empty_string(&mut data);
+    data.put_u32_le(game_id);
+    data.to_vec()
+}
+
+/// Builds a READY_TO_PLAY packet
+/// Format: Empty String [00]
+pub fn build_ready_to_play_packet() -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_empty_string(&mut data);
+    data.to_vec()
+}
+
+/// Builds a SERVER_TO_CLIENT_ACK packet
+/// Format: Empty String [00], 0 (4B), 1 (4B), 2 (4B), 3 (4B)
+pub fn build_server_to_client_ack_packet() -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_empty_string(&mut data);
+    data.put_u32_le(0);
+    data.put_u32_le(1);
+    data.put_u32_le(2);
+    data.put_u32_le(3);
+    data.to_vec()
+}
+
+/// Builds a DROP_GAME packet
+/// Format: Username (NB), Player Number (1B)
+pub fn build_drop_game_packet(username: &str, player_number: u8) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_string_with_null(&mut data, username);
+    data.put_u8(player_number);
+    data.to_vec()
+}
+
+/// Builds a GAME_DATA packet
+/// Format: Empty String [00], Data Length (2B), Game Data (NB)
+pub fn build_game_data_packet(game_data: &[u8]) -> Vec<u8> {
+    let mut data = BytesMut::new();
+    put_empty_string(&mut data);
+    data.put_u16_le(game_data.len() as u16);
+    data.put(game_data);
+    data.to_vec()
+}
+
+/// Builds a GAME_CACHE packet
+/// Format: Empty String [00], Position (1B)
+pub fn build_game_cache_packet(position: u8) -> Vec<u8> {
+    vec![0x00, position]
 }
